renaming icd_lookup.py to linearization_entity.py for clarity
