+++ icd_api/linearization_entity.py
